Update Simmelfolk festival to consolidate organisers (#1018)
A non-profit association was founded to host the Simmelfolk Festival and other Festivals (in Germany).
Thus Simmelfolk festival was updated to reflect this change in organisation.

To make it consistent, also past events were updated (to allow for better filtering and searching).

In this process, also the website was moved and renewed, such that those links were updated too. The website of the new association can be found at https://dckn.de/.

The people behind the Simmelfolk festival are still the same, but the team got a bit bigger (including me now taking care of IT).